Refactor: DTO 클래스 패키지 경로에서 service 제거 #13
