+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/auth/port/AuthFintechQueryPort.java
@@ -1,6 +1,6 @@
 package com.nextdoor.nextdoor.domain.auth.port;
 
-import com.nextdoor.nextdoor.domain.auth.service.dto.AuthFintechQueryDto;
+import com.nextdoor.nextdoor.domain.auth.dto.AuthFintechQueryDto;
 
 public interface AuthFintechQueryPort {
