+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/auth/port/AuthMemberCommandPort.java
@@ -1,7 +1,7 @@
 package com.nextdoor.nextdoor.domain.auth.port;
 
-import com.nextdoor.nextdoor.domain.auth.service.dto.MemberCommandDto;
-import com.nextdoor.nextdoor.domain.auth.service.dto.MemberQueryDto;
+import com.nextdoor.nextdoor.domain.auth.dto.MemberCommandDto;
+import com.nextdoor.nextdoor.domain.auth.dto.MemberQueryDto;
 
 public interface AuthMemberCommandPort {
