+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/auth/port/AuthMemberQueryPort.java
@@ -1,6 +1,6 @@
 package com.nextdoor.nextdoor.domain.auth.port;
 
-import com.nextdoor.nextdoor.domain.auth.service.dto.MemberQueryDto;
+import com.nextdoor.nextdoor.domain.auth.dto.MemberQueryDto;
 
 import java.util.Optional;
