+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/query/AuthMemberQueryAdapter.java
@@ -2,7 +2,7 @@ package com.nextdoor.nextdoor.query;
 
 import com.nextdoor.nextdoor.common.Adapter;
 import com.nextdoor.nextdoor.domain.auth.port.AuthMemberQueryPort;
-import com.nextdoor.nextdoor.domain.auth.service.dto.MemberQueryDto;
+import com.nextdoor.nextdoor.domain.auth.dto.MemberQueryDto;
 import com.nextdoor.nextdoor.domain.member.domain.model.QMember;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
